MATLAB script to test the Attempt 2 model
Tests the second attempt model with the dataset 'dataset_new_test.mat' that is more representative of all the images encountered in the actual game. Was used to determine why the test results were good, but the actual game performance in Attempt 2 was very poor with lots of false positives.